requirements.txt
Flask
Flask-Talisman
Flask-Session
python-dotenv
pycryptodome
pypqc
redis